Clean-up

+++ src/main/java/com/example/devcon/ui/controllers/UserController.java
@@ -1,6 +1,6 @@
 package com.example.devcon.ui.controllers;
 
-import com.example.devcon.common.domain.User;
+import com.example.devcon.users.User;
 import com.example.devcon.users.UserRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.crypto.password.PasswordEncoder;
@@ -54,7 +54,7 @@ public class UserController {
     }
 
     @PostMapping("/update/{id}")
-    public String updateUser(@PathVariable("id") int id, @Valid User user, BindingResult result) {
+    public String updateUser(@PathVariable("id") long id, @Valid User user, BindingResult result) {
         if (result.hasErrors()) {
             user.setId(id);
             return "/users/update";
